change the cast from int to a long in json:len

+++ json/src/main.cpp
@@ -335,7 +335,7 @@ namespace json{
         if (args.size() != 1)
             throw std::runtime_error("json:len: needs 1 argument: jsonAsList or jsonAsObject");
         if (args[0].usertype_ref().as<nlohmann::json>().is_array() || args[0].usertype_ref().as<nlohmann::json>().is_object())
-            return Value(static_cast<int>(args[0].usertype_ref().as<nlohmann::json>().size()));
+            return Value(static_cast<long>(args[0].usertype_ref().as<nlohmann::json>().size()));
         else
             throw std::runtime_error("json:len: json must be a jsonAsList or a jsonAsObject");
     }
